+++ src/wilderness/resource_depletion_simple.h
@@ -12,16 +12,16 @@
 /* ===== FUNCTION PROTOTYPES ===== */
 
 /* Basic depletion functions */
-float get_resource_depletion_level(room_rnum room, int resource_type);
+double get_resource_depletion_level(room_rnum room, int resource_type);
 void apply_harvest_depletion(room_rnum room, int resource_type, int quantity);
 bool should_harvest_fail_due_to_depletion(room_rnum room, int resource_type);
-float get_harvest_success_modifier(room_rnum room, int resource_type);
-const char *get_depletion_level_name(float resource_level);
+double get_harvest_success_modifier(room_rnum room, int resource_type);
+const char *get_depletion_level_name(double resource_level);
 
 /* Conservation functions (stubs for now) */
 void update_conservation_score(struct char_data *ch, int resource_type, bool sustainable);
-float get_player_conservation_score(struct char_data *ch, int resource_type);
-const char *get_conservation_status_name(float score);
+double get_player_conservation_score(struct char_data *ch, int resource_type);
+const char *get_conservation_status_name(double score);
 void show_resource_conservation_status(struct char_data *ch, int x, int y);
 void show_regeneration_analysis(struct char_data *ch, int x, int y);
